Port src/marks/{frame,rule} from JS to TS (stage 1)
Inline the hand-written .d.ts type declarations into single .ts source
files, deleting the original .js and .d.ts pairs. No runtime behaviour
change: parameter and option types now annotate what the JSDoc and
.d.ts already documented.
